change IngredientType of IngredientData
IngredientType is a Enum
By using scriptableObject instead of a enum
we can add more type of ingredient easily
also, as I noted, prefab/icon should be packaged together with ingredient.
So I suggest to use IngredientData, which is a scriptableObject to distinguish different ingredient. Instead of the enum ingredientType.
Also I implement some data from design plan.
Noted that the order system need to retest.